Renaming preview ScanResult
